% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add}
\alias{add}
\alias{add_weights}
\alias{add_residuals}
\alias{add_rstandard}
\alias{add_i2}
\alias{add_h2}
\alias{add_tau2}
\alias{add_blups}
\title{Add model values to a table of meta-analysis results}
\usage{
add_weights(x, .meta = NULL, group = NULL)

add_residuals(x, .meta = NULL)

add_rstandard(x, .meta = NULL)

add_i2(x, .meta = NULL)

add_h2(x, .meta = NULL)

add_tau2(x, .meta = NULL)

add_blups(x, .meta = NULL, exponentiate = FALSE)
}
\arguments{
\item{x}{a tidied meta-analysis}

\item{.meta}{a meta-analysis object. If not `NULL`, then looks for a `meta`
column in `x` containing an overall meta-analysis}

\item{group}{a grouping variable}

\item{exponentiate}{logical. Should the results be exponentiated?}
}
\value{
a `tbl`
}
\description{
Kyle, you got this and the rest of the tidymeta stuff from Malcome
DO NOT FORGET TO CREDIT HIM. Malcom if you're reading this I only did
it as a dirty hack to get your awesome work to work.
}
\details{
`add_weights()` adds study weights to the model. `add_residuals()` and
`add_rstandard()` add residuals. `add_i2()`, `add_h2()`, `add_tau2()`
}
\examples{

library(dplyr)

iud_cxca \%>\%
  group_by(group) \%>\%
  meta_analysis(, yi = lnes, sei = selnes, slab = study_name) \%>\%
  add_i2()

}
