% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_start.R
\name{add_meta}
\alias{add_meta}
\title{Add a meta-analysis to a tbl}
\usage{
add_meta(x, meta, ...)
}
\arguments{
\item{x}{a `data.frame` produced by `tidy()`}

\item{meta}{a meta-analysis object (e.g. `rma.uni`)}

\item{...}{additional arguments to `as.tbl`}
}
\value{
a `tbl` with a `listcol`, `meta`, containing the meta-analysis
}
\description{
`add_meta()` add a meta-analysis object to a tidied meta-analysis.
}
\examples{

library(metafor)
library(broom)

ma <- rma(yi = lnes, sei = selnes, slab = study_name, data = iud_cxca)

ma \%>\%
  tidy() \%>\%
  add_meta(ma)

}
