% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_start.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Meta-Analysis}
\usage{
meta_analysis(.data, .f = metafor::rma, ..., conf.int = TRUE,
  exponentiate = FALSE, include_studies = TRUE, bind_data = FALSE,
  weights = TRUE)
}
\arguments{
\item{.data}{a `data.frame` containing the data for the meta-analysis}

\item{.f}{a meta-analysis function. Default is [metafor::rma]}

\item{...}{additional arguments passed to `.f`}

\item{conf.int}{logical. Include confidence intervals?}

\item{exponentiate}{logical. Should the estimates and (if `conf.int` =
`TRUE`) confidence intervals be exponentiated?}

\item{include_studies}{logical. Include indidiual study coefficients?}

\item{bind_data}{logical. Bind original data to output?}

\item{weights}{logical. Include study weights in output?}
}
\value{
a `tbl`
}
\description{
A wrapper for meta-analytic functions that automatically tidies it and adds
the meta-analysis object to the `tbl` in a `listcol`.
}
\examples{

meta_analysis(iud_cxca, yi = lnes, sei = selnes, slab = study_name)

}
