% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaverse.R
\docType{package}
\name{metaverse-package}
\alias{metaverse}
\alias{metaverse-package}
\title{metaverse}
\description{
metaverse is an awesome package for evidence synthesis in R.
It combines functions from other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmetaverse/metaverse}
  \item Report bugs at \url{https://github.com/rmetaverse/metaverse/issues}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martinjwestgate@gmail.com} (0000-0003-0854-2034)

Authors:
\itemize{
  \item Eliza Grames \email{eliza.grames@uconn.edu} (0000-0003-1743-6815)
  \item Charles Gray \email{charlestigray@gmail.com} (0000-0002-9978-011X)
  \item W. Kyle Hamilton \email{whamilton@ucmerced.edu} (0000-0002-8642-7990)
  \item Emily Kothe \email{emily.kothe@deakin.edu.au} (0000-0003-1210-0554)
  \item Luke McGuinness \email{luke.mcguinness@bristol.ac.uk} (0000-0001-8730-9761)
  \item Rose O'Dea \email{rose.eleanor.o.dea@gmail.com} (0000-0001-8177-5075)
  \item Alfredo Sanchez-Tojar \email{alfredo.tojar@gmail.com} (0000-0002-2886-0649)
  \item Michael Schermann \email{mschermann@scu.edu} (0000-0003-1210-0554)
}

}
