% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaverse_package.R
\docType{package}
\name{metaverse}
\alias{metaverse}
\alias{metaverse-package}
\title{Metaverse: workflows for evidence synthesis}
\description{
Metaverse is a prototype package for evidence synthesis (ES) workflows. It doesn't provide new functions, but rather imports a set of underlying packages that provide a useful set of functions for ES. This version imports the following packages:
\itemize{
  \item{\code{litsearchr}: Quasi-automatic search strategy development for systematic reviews}
  \item{\code{revtools}: Article screening for evidence synthesis using manual or visual methods}
 \item{\code{metaDigitise}: High-throughput, flexible and reproducible extraction of data from figures in primary research papers}
 \item{\code{metadat}: Example datasets for meta-analysis}
 \item{\code{metafor}: Meta-analysis}
  \item{\code{robvis}: Publication quality risk-of-bias figures}
  \item{\code{tidymeta}:  Tidy and plot meta-analyses from popular meta-analytic tools}
}
}
