% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaverse_packages}
\alias{metaverse_packages}
\title{List all packages in the metaverse}
\usage{
metaverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metaverse in the list?}
}
\description{
List all packages in the metaverse
}
\examples{
metaverse_packages()
}
