% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_start.R
\name{pull_meta}
\alias{pull_meta}
\title{Pull the meta-analysis object}
\usage{
pull_meta(x)
}
\arguments{
\item{x}{a tidied meta-analysis}
}
\value{
an object created by a meta-analysis
}
\description{
Pull the meta-analysis object
}
\examples{

meta_analysis(iud_cxca, yi = lnes, sei = selnes, slab = study_name) \%>\%
  pull_meta()
}
