% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_all.R
\name{forest_plot}
\alias{forest_plot}
\alias{influence_plot}
\alias{cumulative_plot}
\alias{funnel_plot}
\title{Quickly plot common visualizations for meta-analyses}
\usage{
forest_plot(x, estimate = estimate, study = study, size = weight,
  shape = type, col = type, xmin = conf.low, xmax = conf.high,
  group = NULL, alpha = 0.75, height = 0, ...)

influence_plot(x, estimate = l1o_estimate, study = study, size = 4,
  shape = 15, col = type, xmin = l1o_conf.low,
  xmax = l1o_conf.high, group = NULL, alpha = 0.75, height = 0,
  sum_lines = TRUE, ...)

cumulative_plot(x, estimate = cumul_estimate, study = study,
  size = 4, shape = 15, col = type, xmin = cumul_conf.low,
  xmax = cumul_conf.high, group = NULL, alpha = 0.75, height = 0,
  sum_lines = TRUE, ...)

funnel_plot(x, estimate = estimate, std.error = std.error, size = 3,
  shape = NULL, col = NULL, alpha = 0.75, reverse_y = TRUE,
  log_summary = FALSE, ...)
}
\arguments{
\item{x}{a tidied meta-analysis}

\item{estimate}{variable name of point estimates}

\item{study}{variable name of study labels}

\item{size}{point size; either an aesthetic variable or a specific shape.}

\item{shape}{shape of the points; either an aesthetic variable or a specific
shape.}

\item{col}{color of the points and lines; either an aesthetic variable or a
specific color.}

\item{xmin}{lower confidence interval variable name}

\item{xmax}{upper confidence interval variable name}

\item{group}{a grouping variable}

\item{alpha}{transparancy level}

\item{height}{line height for error bars}

\item{...}{additional arguments}

\item{sum_lines}{logical. Should vertical lines demarcating the summary
estimate and confidence intervals be included?}

\item{std.error}{variable name of standard error variable}

\item{reverse_y}{logical. Should the y-axis be reversed?}

\item{log_summary}{logical. Should the estimate and confidence intervals be
log-transformed?}
}
\value{
a `ggplot2` object
}
\description{
`forest_plot()` presents study and summary estimates. `influence_plot()`
shows the forest plot of senstivity analyses using `senstivity()`.
`cumulative_plot()` shows the forest plot for `cumulative()`. `funnel_plot()`
plots standard errors against the summary esitimate to assess publication
bias.
}
\details{
These were all taken from Malcolm Barrett, Kyle whatever you do NOT forget to
have Charles talk to him and make sure he is credited 100% for this.
}
\examples{

library(dplyr)

ma <- iud_cxca \%>\%
  group_by(group) \%>\%
  meta_analysis(yi = lnes, sei = selnes, slab = study_name)

forest_plot(ma)

funnel_plot(ma)

ma \%>\%
  sensitivity() \%>\%
  influence_plot()

ma \%>\%
  cumulative() \%>\%
  cumulative_plot()

}
