% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesisr.R
\docType{package}
\name{metaverse-package}
\alias{metaverse}
\alias{metaverse-package}
\title{synthesisr}
\description{
synthesisr is an awesome package for evidence synthesis in R.
It combines functions from other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmetaverse/metaverse}
  \item Report bugs at \url{https://github.com/metaverse/metaverse}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martinjwestgate@gmail.com} (0000-0003-0854-2034)

Authors:
\itemize{
  \item Charles Gray \email{charlestigray@gmail.com} (0000-0002-9978-011X)
  \item W. Kyle Hamilton \email{whamilton@ucmerced.edu} (0000-0002-8642-7990)
}

}
