% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaverse_package.R
\docType{package}
\name{metaverse}
\alias{metaverse}
\title{Metaverse: workflows for evidence synthesis}
\description{
Metaverse is a  package for evidence synthesis workflows. It doesn't provide new functions, but rather imports underlying packages that provide useful functions for evidence synthesis projects such as systematic review and meta-analysis. \code{metaverse} imports the following packages:
\itemize{
  \item{\code{synthesisr}: Import & de-duplication of bibliographic data}
  \item{\code{litsearchr}: Quasi-automatic search strategy development for systematic reviews}
  \item{\code{revtools}: Article screening for evidence synthesis using manual or visual methods}
 \item{\code{metaDigitise}: High-throughput, flexible and reproducible extraction of data from figures in primary research papers}
  \item{\code{robvis}: Publication quality risk-of-bias figures}
 \item{\code{metafor}: Statistical tools for meta-analysis}
}
}
